Speeds up Android by avoiding unnecessarily cycling in the event loop.
With this change, TextInput._onChange can be called (and update _lastNativeText properly)

Without this change, it doesn't get updated. Which leaves the next componentDidUpdate to call setNativeProps({text:...}) again unnecessarily, which (I believe) is triggering a re-layout, and triggering componentDidUpdate, leading to this cycling in the event loop that makes the Android app slower for me.