Fix off-by-one errors.
I'm using a <FlatList horizontal windowSize={1} initialNumToRender={1} initialScrollIndex={1} pagingEnabled /> (where each element is exactly the width of the screen).

The existing calculation would always detect a window size of 2(!). Namely because checking [375, 375, 750, 750] would cause it to match the first element (of 0, 375) and second element (of 375, 750).

This causes the elements to be sized as (0, 374) and (375, 749), and in turn causes our search window to be [375, 375, 749, 749], which in turn would give a window size of exactly one, as expected.

In the case of natural scroll lists, this is essentially irrelevant. In the case of pagingEnabled lists, this can result in an extra frame being rendered, and since I'm dealing with performance issues, I'm trying to minimize over-draw (at least on initial render), and want to be sure it *only* prioritizes the one element on the screen.